Variable for total memory is now uint64_t instead of uint32_t. Same changes reflected to insert_uint_in_str. Added pmm files. Current pmm functions: initialize pmm with total memory and total pages and stub function to initialize a memory region obtained from mmap.

+++ code/src/include/string.h
@@ -18,6 +18,6 @@ extern void *memset(void *__dest, unsigned char __val, size_t __num_bytes);
 //inserts a number in a given string starting in reverse from the specified string index
 //string length must be greater than the index given
 //returns 0 on success, -1 in case of any error
-extern int insert_uint_in_str(char *__str, size_t __num_last_pos, unsigned int __num);
+extern int insert_uint_in_str(char *__str, size_t __num_last_pos, uint64_t __num);
 
 #endif
